backend/requirements.txt
# FastAPI and Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Data Validation
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# CORS and Middleware
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0.post0
